Improved OPTIONS handler.

+++ dataPumper/src/main/java/core/main/DatabasePumper.java
@@ -24,8 +24,8 @@ public abstract class DatabasePumper {
 	
 	protected boolean pureRandom = false;
 	
-	public abstract void pumpDatabase(float percentage, String fromTable);
-	public abstract void pumpDatabase(float percentage);
+	public abstract void pumpDatabase(double percentage, String fromTable);
+	public abstract void pumpDatabase(double percentage);
 	
 	public void setPureRandomGeneration() {
 		pureRandom = true;
