+++ dataPumper/src/main/java/core/main/DatabasePumperOBDA.java
@@ -49,7 +49,7 @@ public class DatabasePumperOBDA extends DatabasePumper {
 		this.dbToPump = dbToPump;
 	}
 	
-	public void pumpDatabase(float percentage, String fromTable){
+	public void pumpDatabase(double percentage, String fromTable){
 		long startTime = System.currentTimeMillis();
 		
 		dbToPump.setForeignCheckOff();
@@ -117,7 +117,7 @@ public class DatabasePumperOBDA extends DatabasePumper {
 	 * @param db
 	 * @param nRows
 	 */
-	public void pumpDatabase(float percentage){
+	public void pumpDatabase(double percentage){
 		
 		long startTime = System.currentTimeMillis();
